fltk: fix build with 10.11 SDK
Use upstream patch for 10.11 SDK, fixing the build on Yosemite when
building with SDKROOT set to Xcode 7.x's unitary 10.11 SDK.

Also, remove the <= Yosemite restriction, since this incidentally
happens to let it build on El Capitan as well. Note that there are still
El Capitan integration issues with the stable spec, which are already
fixed upstream and available in the devel spec.

Closes #1600.

Signed-off-by: ilovezfs <ilovezfs@icloud.com>