perl: include HOMEBREW_PREFIX in INC by default
A lot of Perl users configure things locally, and we nudge them to do so in
the caveats, but our defaults should be a little more sane so users can use
Perl "out of the box" with minimal setup.

At the moment Perl only checks inside the Cellar for other Perl modules:
```
@INC:
    /usr/local/Cellar/perl/5.24.0/lib/perl5/site_perl/5.24.0/darwin-thread-multi-2level
    /usr/local/Cellar/perl/5.24.0/lib/perl5/site_perl/5.24.0
    /usr/local/Cellar/perl/5.24.0/lib/perl5/5.24.0/darwin-thread-multi-2level
    /usr/local/Cellar/perl/5.24.0/lib/perl5/5.24.0
```

To reduce the amount users need to carry custom variables around in their shell
profiles, let's tell Perl to check our prefix by default as well. This removes
the need to alter the environment manually when installing things like
Imagemagick's Perl element against Homebrew's Perl.

Ideally we would overwrite the sitelib, but Perl expects to be able to mkdir & copy
files into that prefix at build and our sandbox breaks that. It'd need a monster
hack to prevent that and handle it post_install, sadly.