elasticsearch: 5.0.0-alpha3 (devel), fix head install (#1693)
* Add devel and fix head install for Elasticsearch

This commit adds a devel install for Elasticsearch, and fixes the head
install for Elasticsearch to reflect changes that have occurred in
master (e.g., changing the name of the plugin script to
"elasticsearch-plugin").

* Use devel? and head? in caveats and tests

This commit simplifies caveats and tests to use devel? and head? instead
of relying on external state to determine which command-line programs
and arguments to use.