docbook2x: remove pointless deps (#1766)
https://github.com/Homebrew/homebrew-core/commit/1075b656d5b10985dfc2439cdf2211ff96b57e93 was just about
a valid commit at the time, but isn't really today. All of these Perl dependencies have
shipped as part of Apple's Perl installation since at least Perl 5.10, which was
added to OS X in 10.6 (Snow Leopard), and 10.6 is the oldest OS X release the core
supports.

Linuxbrew or Tigerbrew may want to look into this more, but for >= 10.6 it's a
completely misleading and never-invoked block.