qjackctl: fix build errors
qmake-qt5 cannot be found by the configure script. This is fixed adding
the prefix of the Qt5 installation as a configure option.
Building with LLVM/Clang raises an error related to a variable sized
array. This is fixed by patching a source file to use a dynamically
allocated array instead.

Fix missing link time dependency for qjackctl

This error was detected by running brew audit --strict --online
qjackctl which showed following error:

     * Formulae are required to declare all linked dependencies.
         Please add all linked dependencies to the formula with:
               depends_on "portaudio" => :linked

Fixes #3278.
Closes #3281.

Signed-off-by: Dominyk Tiller <dominyktiller@gmail.com>