ghc: improve Sierra compatibility
On Sierra, use new upstream dynamic-library-dirs field to avoid

  ghc: panic! (the 'impossible' happened)
  malformed mach-o: load commands size (x) > 32768

where x is some number larger than 32768.

Note that this fix requires a corresponding update to cabal-install and
that `cabal new-build` hasn't been updated to use the new field yet, so
will continue to fail in the same scenarios as it already does.

Also, bump the vendored gmp to 6.1.1 for all macOS versions, and stop
using nm-classic since llvm-nm -P is POSIX compliant now.

Closes #6358.

Signed-off-by: ilovezfs <ilovezfs@icloud.com>