zsh: remove caveat about online help
Setting HELPDIR to #{HOMEBREW_PREFIX}/share/zsh/help is pointless
because it points to #{prefix}/share/zsh/help by default which is even
more reliable than the symlink. Without the Homebrew-specific path (the
only thing that remotely justifies this caveat), it makes no sense to
mention this topic in caveats at all; otherwise there are a million
things of equal or more importance that could be mentioned. Users should
be expected to RTFM:
http://zsh.sourceforge.net/Doc/Release/User-Contributions.html#Accessing-On_002dLine-Help.

Closes #7500.

Signed-off-by: Mike McQuaid <mike@mikemcquaid.com>