jigdo: don't build the GUI
Jigdo upstream is a dead project. It consists of two components: a GUI,
that was broken and unusable before the rest of the project stopped, and
"jigdo-lite" (etc.), CLI tools that still work and have some life support.

Stop building the GUI which is completely useless anyway. This reduces the
dependencies of this formula significantly.

Closes #10034.

Signed-off-by: ilovezfs <ilovezfs@icloud.com>