pilosa: Fix version and time params and brew test
The build parameters for Version and BuildTime were not being passed in
correctly, so the built binary did not contain the version number. Also,
the command invocation used in `brew test` passed an invalid `--bind`
option that wasn't caught by the test.

This patch uses the Makefile to run the build command, which correctly
sets the version and build time, and it removes the invalid `--bind`
option.

Closes #14532.

Signed-off-by: Alex Dunn <dunn.alex@gmail.com>