faas-cli 0.4.11
This commit fixes the build args to set the GitCommt var at build time
after it moved from the `main` to `commands` package.

It also updates the cli syntax used for the test to the new Cobra based
syntax and adds a new test that checks that the version has been set
correctly.

Also bumps the release to `0.4.11`.

Closes #17610.

Signed-off-by: ilovezfs <ilovezfs@icloud.com>