git-gui: split from git formula.
As a tcl-tk dependency for git is not considered acceptable, this change
moves the gitk and git-gui tools to a new formula and sets the NO_TCLTK
flag when building git.

Additionally, this change ensures that we no longer install a .app for
git-gui. Instead, we use the homebrew-installed tcl-tk wish binary to run
git-gui and gitk.

Without this change, all dialogs that are shown by git-gui are blank
windows and can only be closed by memorizing the location of the buttons.

This fixes #36390 and #39987

This is yet another alternative approach to #42788 and #42789.