coredns 1.6.9 (new formula)
CoreDNS developers made the decision to remove the Homebrew Formula that
was previously hosted in their own Homebrew Tap [1] on April 23, 2020.
I find this decision unfortunate and would like to propose the addition
of the Formula to the core library for others to use.

The default settings will proxy all requests to hostnames not found in
your host file to Cloudflare, Google or IBM Threat Safe DNS servers in
that specific order. Users are encouraged to edit the configuration file
before starting the service.

[1] https://github.com/coredns/deployment/commit/f51d10388e85647a7f2f

Closes #55699.

Signed-off-by: Dawid Dziurla <dawidd0811@gmail.com>