uftp: fixes to plist for "brew audit --strict"
I am not sure if this is what has been causing bottling issues
for the formula.  For better or worse the "uftpd" binary definitely
ends up in bin/ not sbin/ so it looks like the audit complaint was
correct.

I haven't actually tried running this program, but it does
build/test/audit ok with this change.

Closes #66378.

Signed-off-by: FX Coudert <fxcoudert@gmail.com>