minizinc: Add dependencies included in regular MiniZinc distribution
This adds dependencies for the solvers that are built into MiniZinc in the
the official MiniZinc bundles. With theses changes these dependencies will provide
internal solving capabilities to the MiniZinc executable. Futhermore, the Gecode
dependency will enable advanced capabilities of the MiniZinc compiler like
compiler optimisation levels and regular expression parsing.

This also changes the test to better reflect the functionality of MiniZinc by
testing its solving capabilities instead of the documentation executable. The
latter does not have a stable interface and its test might break in future versions.

Closes #66630.

Signed-off-by: Sean Molenaar <1484494+SMillerDev@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>