Remove unnecessary livecheck blocks
Historically, livecheck has checked formula URLs in order of `head`,
`stable`, `homepage` when a `livecheck` block isn't present. However,
we prefer to align checks with the `stable` source these days, so we
recently updated the order to check `stable` before `head`.

As such, we can remove many `livecheck` blocks that only contain
`url :stable`. The instances in this commit were simply acting as a
safeguard to prevent the formula from switching to checking the
`head` URL if one was added in the future.