mtools: fix build errors on Apple Silicon
The mtools configure script incorrectly detects the stat64 function, and
this causes build errors. This forces the configure script to assume
that the function doesn't exist.

The pv formula had a similar issue and included the following
information:

  Since macOS 10.6, stat64 variants are equivalent to plain stat, and
  the suffixed versions have been removed in macOS 11 on Apple Silicon.
  See stat(2).

I tested several of the mtools utilities manually to verify the fix and
they seem to work fine.

Signed-off-by: Frank Hunleth <fhunleth@troodon-software.com>