rt-audio: fix shim pollution
This formula was failing to bottle due to shim-pollution checks
becoming stricter since it was last updated:

  * Files were found with references to the Homebrew shims directory.
    The offending files are:
      share/doc/rt-audio/Makefile
      share/rt-audio/tests/teststops
      share/rt-audio/tests/record
      share/rt-audio/tests/playsaw
      share/rt-audio/tests/Makefile
      share/rt-audio/tests/testall
      share/rt-audio/tests/duplex
      share/rt-audio/tests/audioprobe
      share/rt-audio/tests/playraw
      share/rt-audio/tests/apinames

The package's "make install" actually is fine but the formula then
installs extra files:

* For the "doc" directory, things like the html and images directory
are valid, but we don't need the doxygen source or its build infrastructure
(The doxygen-produced files are in the upstream tarball, we don't
even build them as part of the formula build)

* The "rt-audio/tests" is just installed so that "brew test" can
reference these files later.  I don't know if this is a good pattern
in general, but the only file it actually uses is testall.cpp so
at least limit it to just that file.

Closes #70720.

Signed-off-by: Carlo Cabrera <30379873+carlocab@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>