temporal_tables: fix compile against postgresql 13
As-is the formula can't build against our postgresql.
Indeed, since some of the symbols that it dymnamically links
against got renamed the pre-bottled binary isn't likely to
be functional right now either.

I'm not sure what the long-term future of this formula is
(the last tagged release upstream was from 2017 and we already
have introduced patches for PG11 and PG12 support) but I guess
for now the easiest is to stay on the patching train.  The
upstream git repo did get a flurry of commits in April 2020
so there is still hope of an updated version in the future.

Closes #70936.

Signed-off-by: Sean Molenaar <1484494+SMillerDev@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>