imagesnap: remove explicit SDK setting from xcodebuild
This "-sdk" flag no longer seems to work; this meant that this
formula no longer builds from source and didn't get a Big Sur
rebottle.

The flag was originally added to this formula in 2011 to work
around some issue with Xcode 4 (9c4ae8bdfa1066e82aba57e72c4c2884d5ce0c62)
It seems to be the only formula that does this so it seem safe
to remove.

Also mark it as explicit public domain per:
  https://github.com/rharder/imagesnap/blob/master/README.md#a-note-about-public-domain

Closes #71132.

Signed-off-by: Alexander Bayandin <a.bayandin@gmail.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>