pokerstove: build against homebrew's copy of googletest
This package stopped building from source at some point.  The
problem is that it had trouble finding the copy of googletest
that it includes in the tarball, but it's easy to just switch
to building against the homebrew copy.

Note that googletest seems to only be a build-time dependency
(verified via "otool -L" that there are no dependencies on
any related dynamic libraries)

Closes #71151.

Signed-off-by: Carlo Cabrera <30379873+carlocab@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>