gcc: add zstd support
This enables the use of `zstd` instead of `zlib` for LTO data
compression. Compression with `zstd` is 4-8x faster than with `zlib`. [1]

This is, for practical purposes, part of the default GCC configuration,
since GCC's build system will enable `zstd` support when it discovers
that `zstd` is available in the build environment. We pass the
`--with-zstd` flag anyway to make sure it uses a brewed `zstd` for that.

Closes Homebrew/linuxbrew-core#22800.

[1] https://gcc.gnu.org/legacy-ml/gcc/2019-06/msg00185.html

Closes #77985.

Signed-off-by: Sean Molenaar <1484494+SMillerDev@users.noreply.github.com>
Signed-off-by: FX Coudert <fxcoudert@gmail.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>