hadolint: build with brewed GHC
The formula declares a build dependency on GHC, but doesn't actually
make use of it. Instead, `stack` downloads its own GHC at build time. [1]

Let's fix that by using Homebrew GHC instead. This should hopefully
allow us to build this on ARM, and likely also removes the need for the
`gmp` dependency on Linux.

[1] Attempting to bottle this on ARM without the changes here produces
the error:

    I don't know how to install GHC for (OSX,AArch64), please install manually

Closes #79037.

Signed-off-by: Nanda H Krishna <me@nandahkrishna.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>