wxmac 3.1.5
Upstream labels this as a "Development Release", but I believe this is
actually the version we want to package in Homebrew/core. We do the same
for some other formulae (e.g. gnupg).

What upstream mean by "Development Release" is that this is the version
that gets new features. What they call a "stable" is probably more
accurately called a long-term support release. [1] Updates to this
branch are bug-fixes only and not new features.

This appears to be the version vendored by `wxpython`, so we should
hopefully be able to use this in place of the bundled one.

[1] The section on the stable release on their downloads page says that
the API for this version has been stable since 11 November 2013.