go@1.16 1.16.7 (new formula)
One noteworthy departure from the other versioned go formulae is that
I've used the main Go formula to bootstrap, instead of specifying a
bootstrap resource. This should make the formula simpler (and hopefully
doesn't break anything...).

We do do this with `go-boring`, and haven't gotten any complaints about
it so far.

Closes #84073.

Signed-off-by: rui <rui@chenrui.dev>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>