ranger: add missing command 'rifle'
Define an entry for command 'rifle'.

Running the `rifle.py` in `#{libexec}/ranger/ext/rifle.py` doesn't work.
Cause `ranger` is not really installed into the brewed python, but
running as a python file by it. To make sure pkg `ranger` being found
in `sys.path`. We must define a helper under "#{libexec}", which is
the working directory and pkg source `ranger/` is stored there.

Closes #86801.

Signed-off-by: Sean Molenaar <1484494+SMillerDev@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>