suite-sparse: remove `tbb` dependency
SuiteSparse is not compatible with the latest version of TBB. See
DrTimothyAldenDavis/SuiteSparse#72.

We didn't encounter any problems in CI about this because the build just
carried on as if TBB were not installed.

We can use `tbb@2020`, but this causes issues for dependents that
currently depend on `tbb`. Since the formula isn't using the `tbb`
dependency currently, and upstream plans to drop it at some point, let's
just remove the dependency.

Closes #87485.

Signed-off-by: Sean Molenaar <1484494+SMillerDev@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>