dispatch-build-bottle: upload failed bottles (#88191)
This is useful for examining built bottles that somehow failed the tests
for some other reason. (e.g. https://github.com/Homebrew/homebrew-core/issues/87708#issuecomment-951088501)

This change is identical to the code that's already in `tests.yml`:

https://github.com/Homebrew/homebrew-core/blob/bb8737e1573708695744cf5fd1c265bc877d4424/.github/workflows/tests.yml#L300-L315

While we're here, let's also echo the timeout, since this is useful to
know while a job is running.