mapserver: fix ARM build (#89192)
This builds fine on ARM, but fails the test with the error

    ==> /opt/homebrew/Cellar/mapserver/7.6.4_1/bin/mapserv -v
    dyld: Library not loaded: @rpath/libmapserver.2.dylib
      Referenced from: /opt/homebrew/Cellar/mapserver/7.6.4_1/bin/mapserv
      Reason: image not found

This requires dropping the `CMAKE_SKIP_RPATH` flag, but given that we
have limited support for using Homebrew-installed Python libraries with
non-Homebrew Python interpreters, this doesn't seem like a huge loss. In
any case, we probably prefer an ARM bottle to allowing use of the Python
module with a non-Homebrew Python.

While we're here, let's change the `cmake` invocation to match the
output of `brew create --cmake`.