profanity: update `inreplace` comment (#89900)
Upstream have merged the fix in profanity-im/profanity#1617. When that
fix is included in a release, we need to pass the location of `brew` to
`configure` because `brew` won't be in `PATH` for the build.

`configure` still knows to check `/opt/homebrew` and `/usr/local` for
`brew`, so our builds should be fine without this. However, the bottle
is not relocatable so this is needed for anyone building from source in
a non-default prefix.