graphviz: keep `gvmap.sh`
We removed this in 721fb4dbd90cc9c637bb94b31467ecdb8f2fda63 because it
was non-executable but being installed into `bin`. It should be
executable now, so there's no reason to remove it.

This also makes it less awkward that we're installing a `gvmap.sh.1`
without also installing the thing the manpage is for.

I've omitted the revision bump because no one else seems to be looking
for this. Everyone else can get the script at the next version bump.

Closes #89918.

Signed-off-by: Sean Molenaar <1484494+SMillerDev@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>