workflows/tests: limit use of `CI-long-timeout` label (#90718)
* workflows/tests: limit use of `CI-long-timeout` label

The `CI-long-timeout` label is currently overused: it is almost always
added to every PR that takes longer than an hour to run, and we suffer
from long CI queues as a consequence. This defeats the purpose of having
the label in the first place.

Let's try to mitigate this by automatically removing the
`CI-long-timeout` label whenever there are at least two open PRs that
already have the label applied.

I employ a little hack to be able to use our `label-pull-requests`
action, since trying to do this with `actions/github-script` is a little
involved. An alternative to this is to use an external action such as
`andymckay/labeler`. [1] I can also look at adapting our action
properly, but this might take a while since I don't know JavaScript.

[1] https://github.com/andymckay/labeler

* Consolidate timeout label jobs

* Fail long-timeout PRs when there are too many of them

This gets around the issue of requiring a token with write permissions
to pull requests.

* Integrate long PR count into `tap_syntax` job

* Improve error message

Also, extract the maximum number of long-timeout PRs to a variable for
clarity.