jruby: fix brew link issue on Linux
No need for a rev bump, a new set of bottles is fine.

Fixes
Error: The `brew link` step did not complete successfully
The formula built, but is not symlinked into /home/linuxbrew/.linuxbrew
Could not symlink bin/bundle
Target /home/linuxbrew/.linuxbrew/bin/bundle
is a symlink belonging to ruby. You can unlink it:
  brew unlink ruby

To force the link and overwrite all conflicting files:
  brew link --overwrite jruby

To list all files that would be deleted:
  brew link --overwrite --dry-run jruby

Possible conflicting files are:
/home/linuxbrew/.linuxbrew/bin/bundle -> /home/linuxbrew/.linuxbrew/Cellar/ruby/3.0.3/bin/bundle

Closes #90817.

Signed-off-by: Issy Long <me@issyl0.co.uk>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>